# Improve performance of copying HDAs and HDCAs