--- a/lib/galaxy/model/metadata.py
+++ b/lib/galaxy/model/metadata.py
@@ -527,7 +527,10 @@ class FileParameter(MetadataParameter):
             return None
         if isinstance(value, galaxy.model.MetadataFile) or isinstance(value, MetadataTempFile):
             return value
-        return session.query(galaxy.model.MetadataFile).get(value)
+        if isinstance(value, int):
+            return session.query(galaxy.model.MetadataFile).get(value)
+        else:
+            return session.query(galaxy.model.MetadataFile).filter_by(uuid=value).one()
 
     def make_copy(self, value, target_context, source_context):
         value = self.wrap(value, object_session(target_context.parent))
@@ -538,8 +541,11 @@ class FileParameter(MetadataParameter):
             # so this would ultimately get overwritten anyway.
             new_value = galaxy.model.MetadataFile(dataset=target_context.parent, name=self.spec.name)
             object_session(target_context.parent).add(new_value)
-            object_session(target_context.parent).flush()
-            shutil.copy(value.file_name, new_value.file_name)
+            try:
+                shutil.copy(value.file_name, new_value.file_name)
+            except AssertionError:
+                object_session(target_context.parent).flush()
+                shutil.copy(value.file_name, new_value.file_name)
             return self.unwrap(new_value)
         return None
 
@@ -550,6 +556,8 @@ class FileParameter(MetadataParameter):
             # as in extended_metadata mode, so there we just accept MetadataFile.
             # We will only serialize MetadataFile in this mode and not push to the database, so this is OK.
             value = value.id or value
+            if not isinstance(value, int) and object_session(value):
+                value = str(value.uuid)
         return value
 
     def from_external_value(self, value, parent, path_rewriter=None):