@jmchilton 2020-10-26
> referencing MetadataFile objects by uuid when copying them, which means we don't need to flush before making the HDA <-> Metadata association

Wonderful, brilliant. Constantly in awe!